Move CustomPage types to cpage.go
Tidying up - move structs from cloudflare.go into the relevant files.

Addresses #49.
